+++ CHANGELOG.md
@@ -1,11 +1,11 @@
 
 # Changelog
 
-## 0.0.7 (Not Released Yet)
+## 0.0.7
 
 * On retirer l'ancien filtre «problèmes» dans la recherche de transaction.
 * Remplacement du trigger Civirule new_campagnodon_transaction par un trigger de type cron «X jours après».
-* Ajout des conditions CiviRules: Status, 
+* Ajout de la condition CiviRules sur le Status des transactions.
 
 ## 0.0.6